pyproject.toml
[tool.poetry]
name = "finnhub-data-extraction"
version = "0.1.0"
description = "Basic implementation of the Finnhub API"
authors = ["sillians <ihuomacbasil@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pandas = "^2.1.2"
websocket-client = "^1.6.4"
finnhub-python = "^2.4.19"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"
confluent-kafka = "^2.3.0"
pyspark = "^3.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
